нехороший long.....сразу то и не заметил

+++ homeworks/HW1/src/main/java/ru.atom/Util.java
@@ -29,7 +29,7 @@ public class Util {
      * @return the sum of all values.
      */
     public static long sum(int[] values) {
-        int param = 0;
+        long param = 0L;
         for (int i = 0; i < values.length; i++) {
             param = param + values[i];
         }
